requirements.txt
fastapi[all]==0.128.0
uvicorn==0.40.0
jinja2==3.1.6
beautifulsoup4==4.14.3
scipy==1.15.2
numpy==2.2.3
paho-mqtt==2.1.0